backend/scrapers/social_scraper.py
import httpx
from bs4 import BeautifulSoup
from typing import List, Dict, Optional
from loguru import logger
from scrapers.utils import get_headers, rate_limiter, scraper_cache, handle_scraping_error
import json


class SocialScraper:
    """Scraper pour détecter les produits viraux sur TikTok et Pinterest"""
    
    def __init__(self):
        self.headers = get_headers()
    
    async def scrape_tiktok_trending(self, hashtag: str = "tiktokmademebuyit", limit: int = 50) -> List[Dict]:
        """
        Détecter les produits viraux sur TikTok via hashtags
        Note: TikTok nécessite une API officielle ou scraping avancé
        Cette version utilise une approche simplifiée
        """
        cache_key = f"tiktok_trending_{hashtag}"
        cached_data = scraper_cache.get(cache_key)
        
        if cached_data:
            return cached_data
        
        products = []
        
        try:
            # Pour une vraie implémentation, utiliser TikTok API ou Playwright
            # Ici on simule avec des données de démonstration
            
            logger.info(f"Scraping TikTok trending products for #{hashtag}")
            rate_limiter.wait()
            
            # Exemple de produits tendances TikTok (à remplacer par vraie API)
            demo_products = [
                {
                    "nom": "LED Strip Lights - TikTok Viral",
                    "prix": 19.99,
                    "url": "https://www.tiktok.com/tag/ledlights",
                    "source": "tiktok",
                    "categorie": "home_decor",
                    "image_url": "https://example.com/led-lights.jpg",
                    "rating": 4.8,
                    "reviews_count": 15000,
                    "stock_status": "in_stock",
                    "description": f"Viral sur TikTok avec #{hashtag}"
                },
                {
                    "nom": "Portable Blender - TikTok Must Have",
                    "prix": 29.99,
                    "url": "https://www.tiktok.com/tag/portableblender",
                    "source": "tiktok",
                    "categorie": "kitchen",
                    "image_url": "https://example.com/blender.jpg",
                    "rating": 4.7,
                    "reviews_count": 12000,
                    "stock_status": "in_stock",
                    "description": f"Produit viral TikTok #{hashtag}"
                },
                {
                    "nom": "Silicone Face Cleaner",
                    "prix": 12.99,
                    "url": "https://www.tiktok.com/tag/skincare",
                    "source": "tiktok",
                    "categorie": "beauty",
                    "image_url": "https://example.com/face-cleaner.jpg",
                    "rating": 4.9,
                    "reviews_count": 20000,
                    "stock_status": "in_stock",
                    "description": f"TikTok viral beauty product #{hashtag}"
                }
            ]
            
            products = demo_products[:limit]
            
            # Mettre en cache
            scraper_cache.set(cache_key, products)
            
            logger.info(f"Successfully scraped {len(products)} TikTok trending products")
            return products
        
        except Exception as e:
            error = handle_scraping_error(e, "tiktok", f"hashtag:{hashtag}")
            return []
    
    async def scrape_pinterest_trending(self, keyword: str = "trending products", limit: int = 50) -> List[Dict]:
        """
        Détecter les produits viraux sur Pinterest
        """
        cache_key = f"pinterest_trending_{keyword}"
        cached_data = scraper_cache.get(cache_key)
        
        if cached_data:
            return cached_data
        
        products = []
        
        try:
            logger.info(f"Scraping Pinterest trending products for: {keyword}")
            rate_limiter.wait()
            
            # Pinterest nécessite une API ou scraping avancé
            # Simulation avec données de démonstration
            
            demo_products = [
                {
                    "nom": "Aesthetic Room Decor Set",
                    "prix": 34.99,
                    "url": "https://www.pinterest.com/pin/aesthetic-room",
                    "source": "pinterest",
                    "categorie": "home_decor",
                    "image_url": "https://example.com/room-decor.jpg",
                    "rating": 4.6,
                    "reviews_count": 8000,
                    "stock_status": "in_stock",
                    "description": f"Pinterest trending: {keyword}"
                },
                {
                    "nom": "Minimalist Jewelry Set",
                    "prix": 24.99,
                    "url": "https://www.pinterest.com/pin/jewelry",
                    "source": "pinterest",
                    "categorie": "fashion",
                    "image_url": "https://example.com/jewelry.jpg",
                    "rating": 4.8,
                    "reviews_count": 10000,
                    "stock_status": "in_stock",
                    "description": f"Viral sur Pinterest: {keyword}"
                },
                {
                    "nom": "Eco-Friendly Water Bottle",
                    "prix": 18.99,
                    "url": "https://www.pinterest.com/pin/waterbottle",
                    "source": "pinterest",
                    "categorie": "sports",
                    "image_url": "https://example.com/bottle.jpg",
                    "rating": 4.7,
                    "reviews_count": 9500,
                    "stock_status": "in_stock",
                    "description": f"Pinterest must-have: {keyword}"
                }
            ]
            
            products = demo_products[:limit]
            
            # Mettre en cache
            scraper_cache.set(cache_key, products)
            
            logger.info(f"Successfully scraped {len(products)} Pinterest trending products")
            return products
        
        except Exception as e:
            error = handle_scraping_error(e, "pinterest", keyword)
            return []
    
    async def detect_viral_products(self) -> List[Dict]:
        """
        Combiner TikTok et Pinterest pour détecter les produits viraux
        """
        logger.info("Detecting viral products across social platforms")
        
        all_products = []
        
        # TikTok
        tiktok_products = await self.scrape_tiktok_trending(limit=25)
        all_products.extend(tiktok_products)
        
        # Pinterest
        pinterest_products = await self.scrape_pinterest_trending(limit=25)
        all_products.extend(pinterest_products)
        
        logger.info(f"Total viral products detected: {len(all_products)}")
        return all_products


# Instance globale
social_scraper = SocialScraper()
